Marconi_Server guaranteed init all parameters
Marconi server now initializes all parameters. It ignores the settings
of the Marconi on starting up, however.

To be implemented soon will be serialization features to save the
parameters of the marconi for reload on the next startup. The
server will still start with some defaults, but a script can be
immediately run to reset the values to previous ones. ALternatively
the SetInitialValues method can be changed to autoload the parameters
from the last session.