Marconi_Server: save/load settings in registry
Add server settings to save the server data to the registry, and
reload those settings at a latter time. Upon server shutdown, the
current settings are saved in the registry and can be reloaded later.

Location of saves in the registery is /Servers/Marconi\ Server/Settings.
Multiple sets of settings may be saved to different keys.